movement controller ok

+++ Game/Inc/Tickable.hpp
@@ -4,5 +4,5 @@ class Tickable
 {
     public:
     virtual void begin() = 0;
-    virtual void tick(uint32_t delta_time) = 0;
+    virtual void tick(float delta_time) = 0;
 };
